chore(get-current-search-params): explain why we have that file
